requirements.txt
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
pylint-django==2.5.3
drf-nested-routers==0.93.4
black==22.8.0